Move config map util to a different file. Incorporate PR review changes
