+++ workflow/common/util.go
@@ -265,11 +265,11 @@ func ProcessArgs(tmpl *wfv1.Template, args wfv1.ArgumentsProvider, globalParams,
 		}
 
 		if inParam.ValueFrom != nil && inParam.ValueFrom.ConfigMapKeyRef != nil {
-			inParam.Value = wfv1.AnyStringPtr("Generated from configMap")
+			inParam.Value = wfv1.AnyStringPtr("Generated at runtime using config maps")
 			if clientset != nil {
 				configMapValue, err := util.GetConfigMaps(*clientset, namespace, inParam.ValueFrom.ConfigMapKeyRef.Name, inParam.ValueFrom.ConfigMapKeyRef.Key)
 				if err != nil {
-					return nil, errors.Errorf(errors.CodeBadRequest, "inputs.parameters.%s malformed. %s", inParam.Name, err)
+					return nil, errors.Errorf(errors.CodeBadRequest, "inputs.parameters.%s malformed: %w", inParam.Name, err)
 				}
 				inParam.Value = wfv1.AnyStringPtr(configMapValue)
 			}
